Consolidate spot+futures strategy registries into a single source of truth
Every shared strategy lived twice — once in `shared_strategies/spot/strategies.py`
and once in `shared_strategies/futures/strategies.py` — with ~750 lines of
copy-pasted implementations. Any bug fix had to land in both files or the
registries silently diverged, with no mechanism to detect drift.

Introduces `shared_strategies/registry.py` as the single source of truth.
Strategies register once with a `platforms=(...)` tuple and optional
`variants` dict for per-platform description / default-param overrides
(used by `momentum`, `rsi`, `macd`, `mean_reversion`). The spot and futures
`strategies.py` files become thin shims (60–65 lines) that materialize a
platform-filtered view via `build_registry(platform)`.

- `--list-json` output for both platforms is byte-identical to main.
- All existing tests pass unchanged (256 tests in shared_strategies/ +
  backtest/tests/test_registry_loader.py).
- New `shared_strategies/test_registry_parity.py` enforces invariants:
  non-empty platforms subset of {spot,futures}, variants keys subset of
  platforms, no double-registration, PLATFORM_ORDER matches platform tags,
  every shim strategy applies cleanly on synthetic data.
- `spot.STRATEGY_REGISTRY is not fut.STRATEGY_REGISTRY` preserved (each
  shim loads registry.py via `spec_from_file_location` under a distinct
  module name).
- Go build + `go-trader init --json` smoke tests for spot / perps / futures
  still produce the expected configs.
- CLAUDE.md "Key Patterns" updated with the new single-source-of-truth flow.

Closes #333

---
