+++ CLAUDE.md
@@ -101,9 +101,9 @@
 - `cfg.AutoUpdate` — `"off"` (default), `"daily"` (once/day), `"heartbeat"` (every cycle); handled in main.go loop + startup; uses `dailyCycles = (24*3600)/tickSeconds`
 - Strategy registry imports: `check_strategy.py` imports from `shared_strategies/spot/strategies.py`; `check_hyperliquid.py`, `check_topstep.py`, and `check_okx.py` (swap mode) import from `shared_strategies/futures/strategies.py` — a new strategy must be registered in both if it needs to work across platforms
 - Adding a cross-platform strategy: create core logic in `shared_strategies/<name>.py` (see `chart_patterns.py`, `liquidity_sweeps.py`), then import+register in both `spot/strategies.py` and `futures/strategies.py`; thin wrapper: `@register_strategy(...)` + `def x(df, **params): return x_core(df, **params)`
-- Adding a new spot/futures strategy (no new platform): (1) add `@register_strategy` function to `shared_strategies/spot/strategies.py`, (2) add same to `shared_strategies/futures/strategies.py`, (3) add short name to `knownShortNames` in `scheduler/init.go`, (4) add param grid entry to `DEFAULT_PARAM_RANGES` in `backtest/optimizer.py` or CI's `test_param_ranges_cover_every_registered_strategy` will fail — auto-discovery handles all platform configs
-- Spot and futures have independent `STRATEGY_REGISTRY` dicts — a new strategy must be added to both files with `@register_strategy` decorator; perps auto-discovers from futures via `discoverStrategies()`
-- Futures registry is a superset of spot (minus `pairs_spread`) and includes perps-only strategies like `delta_neutral_funding`; perps derives directly from futures discovery
+- Adding a new spot/futures strategy (no new platform): (1) add `@register` function to `shared_strategies/registry.py` with appropriate `platforms=(...)` tuple (default `("spot","futures")`); (2) append the name to the matching list(s) in `PLATFORM_ORDER` at the bottom of `registry.py` (controls `--list-json` order — keep this byte-stable for agent tooling); (3) if spot and futures flavors differ in description or defaults (see `momentum`, `rsi`, `macd`, `mean_reversion`), pass `variants={"futures": {"description": ..., "default_params": {...}}}` rather than duplicating the function; (4) add short name to `knownShortNames` in `scheduler/init.go`; (5) add param grid entry to `DEFAULT_PARAM_RANGES` in `backtest/optimizer.py` or CI's `test_param_ranges_cover_every_registered_strategy` will fail — auto-discovery handles all platform configs. `shared_strategies/spot/strategies.py` and `shared_strategies/futures/strategies.py` are thin shims that materialize a platform-filtered view via `build_registry(platform)` — **do not edit them to add strategies**
+- Single source of truth: every strategy implementation lives exactly once in `shared_strategies/registry.py`. Spot-only strategies (`pairs_spread`) set `platforms=("spot",)`; futures-only strategies (`breakout`, `delta_neutral_funding`, `triple_ema_bidir`) set `platforms=("futures",)`. `shared_strategies/test_registry_parity.py` enforces invariants (no double-registration, variants subset of platforms, PLATFORM_ORDER matches platform tags, every shim strategy applies cleanly on synthetic data)
+- Spot and futures shims each build a **fresh** `STRATEGY_REGISTRY` dict via `_load_registry_module()` + `build_registry()` — they're not the same object (regression test: `spot.STRATEGY_REGISTRY is not fut.STRATEGY_REGISTRY`). Perps auto-discovers from futures via `discoverStrategies()` in `scheduler/init.go`
 - New strategies also need: (1) `knownShortNames` entry in `init.go` for the `"name": "abbrev"` mapping, (2) `defaultSpotStrategies` / `defaultPerpsStrategies` / `defaultFuturesStrategies` fallback entries in `init.go`
 - Strategy discovery: `shared_strategies/spot/strategies.py --list-json`, `shared_strategies/options/strategies.py --list-json`, and `shared_strategies/futures/strategies.py --list-json` output JSON arrays of `{"id":..., "description":...}`
 - `apply_strategy(name, df, params)` — optional `params` dict merges with strategy defaults; used to inject external data (e.g. funding rates) into strategies that need non-OHLCV inputs
